moved Header, NavMaximized, and NavMinimized home components to global components folder
