+++ src/components/NavMinimized/index.jsx
