+++ src/components/index.js
@@ -1,7 +1,15 @@
 // components
 
+import Header from './Header';
 import SignInButton from './SignInButton';
+import NavMaximized from './NavMaximized';
+import NavMinimized from './NavMinimized';
 
 
 
-export { SignInButton };
+export {
+  Header,
+  SignInButton,
+  NavMaximized,
+  NavMinimized
+};
